selecttest/Selection-getRangeAt.html

<!DOCTYPE html>
<title>The getRangeAt method</title>
<div id=log></div>
<script src=support/testharness.js></script>
<script src=support/testharnessreport.js></script>
<script>
test(function() {
  var sel = getSelection();
  var range = document.createRange();
  sel.addRange(range);
  assert_throws("INDEX_SIZE_ERR", function() { sel.getRangeAt(-1); })
  assert_throws("INDEX_SIZE_ERR", function() { sel.getRangeAt(1); })
});
</script>
